requirements.txt
streamlit==1.27.2
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.0.1
matplotlib==3.8.0